+++ changelog.md
@@ -7,6 +7,7 @@
 * Fixed aircraft control animations
 * Fixed overlay on fabric being funky with some mods
 * Aircraft now take damage and slow down in lava
+* Fixed Bamboo Hopper floating on modded water-like fluids
 
 # 1.4.1